authicuno/requirements.txt
discord.py
mysql-connector-python
sqlalchemy
pytest